Replace session cookies with an authorization header.
We previously used session cookies with express-session, because this
was the simplest possible way of simulating some notion of persistent
session. However, GraphQL playground at some point changed its default
behavior of using fetch with the `credentials: 'include'` option which
adds cookies to fetch requests, making the session id change for every
request. This makes manual testing in the playground unnecessarily
difficult. The app just generates a random user id every time it is
launched, so that mutations can be tested easily.